malam ini

+++ database/migrations/2023_10_05_00012_create_t2024_lokasis_table.php
@@ -37,7 +37,7 @@ return new class extends Migration
 			$table->string('tanam_pict')->nullable();
 			$table->date('tgl_panen')->nullable();
 			$table->date('tgl_akhir_panen')->nullable();
-			$table->decimal('volume')->nullable();
+			$table->double('volume')->nullable();
 			$table->string('panen_doc')->nullable();
 			$table->string('panen_pict')->nullable();
 			$table->integer('status')->nullable();
